Allow quickstart compose file to be used more flexibly
The quickstart `docker-compose.yml` file hardcodes some parameters. Docker
Compose supports an interpolation mechanism which allows the same
defaulting to happen, but in a way which permits the user to modify
these parameters when running docker-compose.

Making this change will allow this Compose file to be used by more
consumers, whilst still enabling the same low-friction quickstart
workflow.

Alternative mechanisms, such as adding a `docker-compose.env` file and
wget'ing it as part of the quickstart, are possible - and don't increase
the quickstart friction very much. This commit implements the mechanism
which doesn't change the quickstart experience *at all*.