Quickstart's local users can have plaintext passwords specified
This commit changes the prehashed password for the `test` user in the
quickstart `docker-compose.yml` file to the same password, but in
plaintext.

This makes it clearer for consumers of this file as to how they
should change the default user and password. Suggesting in the status
quo that a hash must be specified, especially one which will contain
dollar signs, makes it less likely a user will override this setting
using the functionality introduced in `7f2e0a2`.