Updated the Nodejs example
Using task inputs and outputs its possible to separate the npm install stage from the npm test stage.
This is useful as it allows the testing stage to run and hopefully only return error messages to do with testing and not something somewhat unrelated.
As this is an example it believe it should not resort to using exec commands as it skips over some of the more interesting features of concourse.

Signed-off-by: Alex Kennedy <alexander.kennedy@parlamind.com>