update task config docs to reflect how Path currently works
concourse/concourse#6597 changed how the `path` works in Concourse and
allowed users to specify an absolute path. The exact commit is this one:

https://github.com/concourse/concourse/commit/2ba0164e8be66100e24c727bd51f04c625398231

This has been the current behaviour since 7.5.0 so I don't think there's
any going back on this behaviour now. I think it may make caching stuff
a lot easier for folks to work with. I'm generally in favour of this
change which I think was kinda snuck in. It was a massive PR.

Signed-off-by: Taylor Silva <dev@taydev.net>